chore: add promptEmail to prompt.js
